fix: audit fixes — row shuffle in watchlist, DB lock, scroll/bg polish, single-instance exit, transparent bars/tables

+++ core/tracker.py
@@ -26,7 +26,7 @@ class _SessionState:
 @dataclass
 class _TickContext:
     now: datetime
-    delta: int
+    delta: float
     slept: bool
     idle_now: float
     focused_name: Optional[str]
